Add Refresh Token, email reset password, test coverage and refactoring

+++ src/modules/cars/useCases/deleteCar/DeleteCarUseCase.ts
@@ -2,6 +2,7 @@ import { inject, injectable } from "tsyringe";
 
 import { DeleteResult } from "typeorm";
 import { ICarsRepository } from "@modules/cars/repositories/ICarsRepository";
+import { IRentalsRepository } from "@modules/rentals/repositories/IRentalsRepository";
 import { CustomError } from "@shared/errors/CustomError";
 
 interface IRequest {
@@ -13,6 +14,9 @@ class DeleteCarUseCase {
   constructor(
     @inject("CarsRepository")
     private carsRepository: ICarsRepository,
+
+    @inject("RentalsRepository")
+    private rentalsRepository: IRentalsRepository,
   ) {}
 
   async execute({ license_plate }: IRequest): Promise<DeleteResult> {
@@ -20,6 +24,11 @@ class DeleteCarUseCase {
 
     if (!carAlreadyExists) {
       throw new CustomError(404, "Car do not exists!");
+    }
+
+    const rentalExists = await this.rentalsRepository.findRentalByCar(carAlreadyExists.id);
+    if (rentalExists) {
+      throw new CustomError(400, "Cannot delete a car with a rent assigned!");
     }
 
     return this.carsRepository.delete(license_plate);
